bcrypt: Move macOS support to a new file.
