+++ dlls/bcrypt/Makefile.in
@@ -6,6 +6,7 @@ EXTRAINCL = $(GNUTLS_CFLAGS)
 C_SRCS = \
 	bcrypt_main.c \
 	gnutls.c \
+	macos.c \
 	md2.c \
 	sha256.c \
 	sha512.c
